Add more expenses and style the right sidebar
Added more expenses under the 'Today' section as well as added a new
section underneath, for 23rd of March 2023.
Created another sidebar on the right side of the screen, where total
expenses for each category are displayed with bars based on the
percentage of the total amount of money spent.